Refactor(Wolfx): Relocate file
