+++ src/pages/scripts/ydits-web/services/api/api.js
@@ -13,7 +13,7 @@ import { Service } from "../../../packages/app-creator/src/service.js";
 import { YahooKmoni } from "./yahoo-kmoni.js";
 import { P2pquake } from "./p2pquake.js";
 import { Dmdata } from "./dmdata.js";
-import { Wolfx } from "./wolfx.js";
+import { Wolfx } from "../wolfx/wolfx.js";
 
 /**
  * APIを扱う
